pyproject.toml
[tool.poetry]
name = "AuthZilla"
version = "0.1.0"
description = ""
authors = ["errbufferoverfl <merrymet@errbufferoverfl.me>"]
readme = "README.md"

[tool.poetry.scripts]
corezilla = "corezilla.run:app"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
sqlalchemy = "^2.0.4"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
flask-marshmallow = "^1.2.1"
marshmallow = "^3.19.0"
marshmallow-enum = "^1.5.1"
marshmallow-sqlalchemy = "^1.1.0"
flask-smorest = "^0.44.0"
flask-security-too = "^5.1.1"
bcrypt = "^4.0.1"
python3-xid = "^1.0.5"
setuptools = "^75.1.0"
requests = "^2.32.3"
argon2-cffi = "^23.1.0"
pyjwt = "^2.10.1"
cryptography = "^44.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
